cleaned and made auth into one servlet
