--- src/main/java/com/DevSync/Servlets/Auth/LoginServlet.java
@@ -1,37 +1,0 @@
-package com.DevSync.Servlets.Auth;
-
-import com.DevSync.Controllers.UtilisateurController;
-import jakarta.inject.Inject;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.annotation.WebServlet;
-import jakarta.servlet.http.HttpServlet;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-
-import java.io.IOException;
-
-@WebServlet("/login")
-public class LoginServlet extends HttpServlet {
-    @Inject
-    private UtilisateurController utilisateurController;
-
-    @Override
-    protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        request.setAttribute("contentPage", "/WEB-INF/Views/Auth/login.jsp");
-        request.getRequestDispatcher("/WEB-INF/app.jsp").forward(request, response);
-    }
-
-    @Override
-    protected void doPost(HttpServletRequest request, HttpServletResponse response) throws IOException, ServletException {
-        String username = request.getParameter("username");
-        String password = request.getParameter("password");
-
-        if (utilisateurController.login(username, password)) {
-            request.setAttribute("successMessage", "Login Successful!");
-            response.sendRedirect(request.getContextPath()+"/home");
-        } else {
-            request.setAttribute("errorMessage", "Invalid username or password.");
-            request.getRequestDispatcher("/WEB-INF/Views/Auth/login.jsp").forward(request, response);
-        }
-    }
-}
