--- src/main/java/com/DevSync/Servlets/Auth/LogoutServlet.java
@@ -1,28 +1,0 @@
-package com.DevSync.Servlets.Auth;
-
-import com.DevSync.Controllers.UtilisateurController;
-import jakarta.inject.Inject;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.annotation.WebServlet;
-import jakarta.servlet.http.HttpServlet;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import jakarta.servlet.http.HttpSession;
-
-import java.io.IOException;
-
-@WebServlet("/logout")
-public class LogoutServlet extends HttpServlet {
-    @Inject
-    private UtilisateurController utilisateurController;
-
-    @Override
-    protected void doGet(HttpServletRequest request, HttpServletResponse response) throws IOException {
-        HttpSession session = request.getSession(false);
-        if (session != null) {
-            session.invalidate();
-        }
-
-        response.sendRedirect(request.getContextPath() + "/login");
-    }
-}
