+++ src/main/webapp/WEB-INF/Views/Components/_navbar.jsp
@@ -29,13 +29,13 @@
                     if (session.getAttribute("username") != null) {
                 %>
                 <li>
-                    <a href="<%= request.getContextPath() %>/logout" class="block py-2 px-3 text-red-900 rounded hover:bg-red-100 md:hover:bg-transparent md:border-0 md:hover:text-red-700 md:p-0 dark:text-white md:dark:hover:text-red-500 dark:hover:bg-red-700 dark:hover:text-white md:dark:hover:bg-transparent">Logout</a>
+                    <a href="<%= request.getContextPath() %>/auth/logout" class="block py-2 px-3 text-red-900 rounded hover:bg-red-100 md:hover:bg-transparent md:border-0 md:hover:text-red-700 md:p-0 dark:text-white md:dark:hover:text-red-500 dark:hover:bg-red-700 dark:hover:text-white md:dark:hover:bg-transparent">Logout</a>
                 </li>
                 <%
                     } else {
                 %>
                 <li>
-                    <a href="<%= request.getContextPath() %>/login" class="block py-2 px-3 text-gray-900 rounded hover:bg-gray-100 md:hover:bg-transparent md:border-0 md:hover:text-blue-700 md:p-0 dark:text-white md:dark:hover:text-blue-500 dark:hover:bg-gray-700 dark:hover:text-white md:dark:hover:bg-transparent">Login</a>
+                    <a href="<%= request.getContextPath() %>/auth/login" class="block py-2 px-3 text-gray-900 rounded hover:bg-gray-100 md:hover:bg-transparent md:border-0 md:hover:text-blue-700 md:p-0 dark:text-white md:dark:hover:text-blue-500 dark:hover:bg-gray-700 dark:hover:text-white md:dark:hover:bg-transparent">Login</a>
                 </li>
                 <%
                     }
